separate tests of time formatter: it is not a Formatter plugin
